lesson 15 draft 2

+++ 15_DatabasesAdvanced/djmarketplace/app_users/forms.py
@@ -1,7 +1,6 @@
 from django import forms
 from django.contrib.auth.forms import UserCreationForm
 from django.contrib.auth.models import User
-from .models import Profile
 
 
 class RegisterForm(UserCreationForm):
@@ -14,4 +13,4 @@ class RegisterForm(UserCreationForm):
 
 
 class ProfileForm(forms.Form):
-    add_to_balance = forms.IntegerField(help_text='добавить к балансу')
+    add_to_balance = forms.DecimalField(help_text='добавить к балансу')
